Correctly implement /etc/meow.conf fallback
I mean, at least, make it correctly work.
Previously, if you only put a configuration file in `/etc/meow.conf`, meowfetch will first panic due to not finding a configuration in the second looked directory, at the root of the home directory, without checking first the fallback directory.
I've literally never programmed in go, but I know some basis in programming, and manage to came with a fix, by adding another variable as `fallback_linux` (dir, Exist & ConfPath) which will check for its existence.
One downfall of that, is that it doesn't use the runtime module, but after trying it a bit, it seemed to also display the architecture, which might or not cause it to not be used properly, but again, I'm a total beginner, so this might be myself who don't know properly how to use those.
Either way, if you really want to not use another bunch of variables, that's totally fine. But be aware that meowfetch doesn't check for it currently.